backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.36
alembic==1.13.0
geoalchemy2==0.14.3
psycopg[binary]==3.2.4
pydantic>=2.10.6,<3
pydantic-settings>=2.7.0,<3
httpx==0.26.0
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
anthropic==0.21.0
google-generativeai==0.4.0
sec-edgar-downloader==5.0.1
